refactoring duplicate functions
