+++ src/Illuminate/Broadcasting/Broadcasters/LogBroadcaster.php
@@ -5,7 +5,7 @@ namespace Illuminate\Broadcasting\Broadcasters;
 use Psr\Log\LoggerInterface;
 use Illuminate\Contracts\Broadcasting\Broadcaster;
 
-class LogBroadcaster implements Broadcaster
+class LogBroadcaster extends AbstractBroadcaster implements Broadcaster
 {
     /**
      * The logger implementation.
